requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
httpx==0.27.0
groq==0.4.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pdfplumber==0.11.4
requests==2.31.0
duckduckgo-search==3.9.6
numpy==1.24.3
torch==2.1.0
transformers==4.35.0
python-dotenv==1.0.0
gunicorn>=22.0.0
gevent==23.9.1






